+++ tests/testthat/test-objects.R
@@ -19,10 +19,22 @@ test_that("create patient", {
 test_that(".process_io_parenteral_vector", {
   
   # Perfect IV sequence
-  expect_equal(.parenteral_vector_process(unlist(strsplit("11111111111111111111", "")), 5), list(c(1, 20)))
+  expect_equal(.parenteral_vector_process(unlist(strsplit("11111111111111111111", "")), 5), list(c(0, 19, NA)))
+  # Perfect IV sequence before gobbledigook
+  expect_equal(.parenteral_vector_process(unlist(strsplit("111111111111111111110011001011", "")), 5), list(c(0, 29, NA)))
+  expect_equal(.parenteral_vector_process(unlist(strsplit("111111111111111111110011001010", "")), 5), list(c(0, 29, NA)))
+  # Perfect IV sequence after gobbledigook
+  expect_equal(.parenteral_vector_process(unlist(strsplit("001100101011111111111111111111", "")), 5), list(c(10, 29, NA)))
+  
   # Perfect IVPO sequence
-  expect_equal(.parenteral_vector_process(unlist(strsplit("11111111110000000000", "")), 5), list(c(1, 20)))
-  expect_equal(.parenteral_vector_process(unlist(strsplit("00000000011111100000", "")), 4), list(c(10, 20)))
+  expect_equal(.parenteral_vector_process(unlist(strsplit("11111111110000000000", "")), 5), list(c(0, 19, 10)))
+  expect_equal(.parenteral_vector_process(unlist(strsplit("00000000011111100000", "")), 4), list(c(9, 19, 15)))
+  # Perfect IVPO sequence before gobbledigook
+  expect_equal(.parenteral_vector_process(unlist(strsplit("11111111110000000000110011001010", "")), 5), list(c(0, 19, 10)))
+  expect_equal(.parenteral_vector_process(unlist(strsplit("00000000011111100000110011001010", "")), 4), list(c(9, 19, 15)))
+  # Perfect IVPO sequence after gobbledigook
+  expect_equal(.parenteral_vector_process(unlist(strsplit("001100101011111111110000000000", "")), 5), list(c(10, 29, 20)))
+  expect_equal(.parenteral_vector_process(unlist(strsplit("001100101000000000011111100000", "")), 4), list(c(19, 29, 25)))
   
   # No sequence
   expect_equal(.parenteral_vector_process(unlist(strsplit("000000000000", "")), 4), list())
@@ -31,13 +43,14 @@ test_that(".process_io_parenteral_vector", {
   
   # One needs to start IV for at least 6 hours uninterrupted.
   expect_equal(.parenteral_vector_process(unlist(strsplit("111110001111011", "")), 4), list())
-  expect_equal(.parenteral_vector_process(unlist(strsplit("111111001111011", "")), 4), list(c(1, 15)))
+  expect_equal(.parenteral_vector_process(unlist(strsplit("111111001111011", "")), 4), list(c(0, 14, NA)))
   
   # Impact of increasing tolerance from 4 to 5
-  expect_equal(.parenteral_vector_process(unlist(strsplit("11111100000111111011", "")), 4), list(c(1, 11), c(12, 20)))
-  expect_equal(.parenteral_vector_process(unlist(strsplit("11111100000111111011", "")), 5), list(c(1, 20)))
-  expect_equal(.parenteral_vector_process(unlist(strsplit("111111#####111111011", "")), 4), list(c(1, 11), c(12, 20)))
-  expect_equal(.parenteral_vector_process(unlist(strsplit("111111#####111111011", "")), 5), list(c(1, 20)))
-  
+  expect_equal(.parenteral_vector_process(unlist(strsplit("11111100000111111011", "")), 4), list(c(0, 10, 6), 
+                                                                                                 c(11, 19, NA)))
+  expect_equal(.parenteral_vector_process(unlist(strsplit("11111100000111111011", "")), 5), list(c(0, 19, NA)))
+  expect_equal(.parenteral_vector_process(unlist(strsplit("111111#####111111011", "")), 4), list(c(0, 5, NA), 
+                                                                                                 c(11, 19, NA)))
+  expect_equal(.parenteral_vector_process(unlist(strsplit("111111#####111111011", "")), 5), list(c(0, 19, NA)))
 })