+++ NEWS.md
@@ -3,12 +3,15 @@
 
 ## Features
 
-* introduce objects
+**ADD HYPERLINKS BEFORE SUBMISSION**
+
+* introduce S4 classes `Patient`, `MedicationRequest`, `TherapyEpisode`
+* introduce therapy methods `get_therapy_table()`, `parenteral_changes_get()`
 
 ## Improvements
 
 * remove dependencies
-* improve handling of time data by SQLite
+* improve handling of time data types by SQLite
 
 ## Documentation
 
@@ -16,7 +19,7 @@
 
 # Ramses 0.2.1
 
-## Bug 
+## Bug fixes
 
 * Corrected incorrect transitive closure settings
 * Corrected prescription edge classification